src/requirements-all.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blis==0.7.11
cachetools==5.2.1
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
configparser==5.2.0
cycler==0.12.1
cymem==2.0.8
datasets==2.8.0
dill==0.3.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
filelock==3.13.1
fonttools==4.49.0
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gensim==4.1.2
h5py==3.10.0
huggingface-hub==0.21.3
hyperopt==0.2.7
idna==3.6
interlap==0.2.7
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
langdetect==1.0.9
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.5.3
msgpack==1.0.8
msgpack-numpy==0.4.8
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.2
pandas==1.4.4
parse==1.20.1
pathlib-abc==0.1.1
pathy==0.11.0
patool==1.15.0
pillow==10.2.0
preshed==3.0.9
protobuf==3.20.3
py4j==0.10.9.7
pyarrow==15.0.0
pydantic==1.8.2
pyparsing==3.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
responses==0.18.0
scikit-learn==1.1.3
scipy==1.8.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
spacy==3.2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
tabulate==0.8.10
thinc==8.0.17
threadpoolctl==3.3.0
tokenizers==0.12.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.2
transformers==4.19.4
typer==0.4.2
typing-extensions==4.5.0
typing-inspect==0.8.0
tzdata==2024.1
urllib3==1.26.6
wasabi==0.10.1
wcwidth==0.2.13
wikitextparser==0.55.8
wrapt==1.16.0
XlsxWriter==3.0.9
xxhash==3.4.1
yarl==1.9.4
zensols.datdesc==0.0.2
zensols.db==1.1.0
zensols.deeplearn==1.6.1
zensols.deepnlp==1.8.0
zensols.install==0.2.2
zensols.nlp==1.6.0
zensols.util==1.13.3
